pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "argos-analytics"
version = "2.0.0"
authors = [{name = "Denisse Alfaro", email = "denisse.alfaper@gmail.com"}]
description = "Signal Analysis and Graph Intelligence for Public Procurement Risk Detection"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pandas>=2.0.0",
    "neo4j>=5.0.0",
    "requests>=2.28.0",
    "tqdm>=4.65.0",
    "python-dotenv>=1.0.0",
    "streamlit>=1.20.0",
    "plotly>=5.13.0",
    "seaborn>=0.13.2",
    "squarify>=0.4.4",
    "networkx>=3.4.2",
    "jinja2>=3.1",
]

[tool.setuptools.packages.find]
where = ["src"]

[dependency-groups]
dev = [
    "ipykernel>=7.2.0",
    "matplotlib>=3.10.8",
]
